// web/src/shared/api/generated/model/gradeReviewSubjectResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface GradeReviewSubjectResult {
  subjectId: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  essayCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  mathStepsCount: number | string;
}